Working on Dijsktra

+++ src/model/VertexConected.java
@@ -3,19 +3,19 @@ package model;
 public class VertexConected<V> implements Comparable<VertexConected <V>> {
 	private Vertex<V> vertex;
 	private Vertex<V> vertexEnd;
-	private double weigth;
+	private int weigth; 
 
 	public Vertex<V> getV() {
 		return vertex;
 	}
 	
-	public double getWeigth() {
+	public int getWeigth() {
 		return weigth;
 	}
 
 	@Override
 	public int compareTo(VertexConected<V> o) {
-		return (int) (weigth- o.getWeigth());
+		return (weigth- o.getWeigth());
 	}
 
 	public Vertex<V> getVertexEnd() {
